src/worker/requirements.txt
dnspython3==1.12.0
pystalkd==1.2.1